Moved email-address-set property.
